Refactor: Use `String.toUri()` and Update Landscapist
This commit includes two main changes:

-   **`NavigationUtils.kt`:**
    -   Updated the `CustomTabsIntent.Builder.build()` extension function to use the `String.toUri()` extension function from `androidx.core.net.toUri()` instead of `Uri.parse()` for creating the referrer URI. This is a more idiomatic Kotlin approach.
-   **Dependency Update:**
    -   Updated the `landscapist` image loading library from version `2.4.8` to `2.4.9`.
    - moving all composables from BaseMainActivity.kt to HomeNav.kt
