+++ UIViews/src/main/java/com/programmersbox/uiviews/BaseMainActivity.kt
@@ -1,191 +1,46 @@
 package com.programmersbox.uiviews
 
-import android.Manifest
 import android.app.assist.AssistContent
-import android.content.Intent
-import android.os.Build
 import android.os.Bundle
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.ExperimentalSharedTransitionApi
-import androidx.compose.animation.SharedTransitionLayout
-import androidx.compose.animation.core.AnimationSpec
-import androidx.compose.animation.core.AnimationState
-import androidx.compose.animation.core.DecayAnimationSpec
-import androidx.compose.animation.core.animateDecay
-import androidx.compose.animation.core.animateTo
-import androidx.compose.animation.expandHorizontally
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.shrinkHorizontally
-import androidx.compose.animation.shrinkVertically
-import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.togetherWith
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.gestures.Orientation
-import androidx.compose.foundation.gestures.draggable
-import androidx.compose.foundation.gestures.rememberDraggableState
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.navigationBars
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.filled.BrowseGallery
-import androidx.compose.material.icons.filled.Extension
-import androidx.compose.material.icons.filled.History
-import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.Star
-import androidx.compose.material.icons.outlined.History
-import androidx.compose.material.icons.outlined.Settings
-import androidx.compose.material3.Badge
-import androidx.compose.material3.BadgedBox
-import androidx.compose.material3.BottomAppBarScrollBehavior
-import androidx.compose.material3.BottomAppBarState
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarDefaults
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.NavigationBarItemDefaults
-import androidx.compose.material3.NavigationRail
-import androidx.compose.material3.NavigationRailItem
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.navigation3.rememberListDetailSceneStrategy
-import androidx.compose.material3.contentColorFor
-import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
-import androidx.compose.material3.windowsizeclass.WindowSizeClass
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
-import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.layout
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.Velocity
-import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.FragmentActivity
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDecorator
-import androidx.navigation.NavDestination
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation3.runtime.NavBackStack
-import androidx.navigation3.runtime.rememberNavBackStack
-import androidx.navigation3.runtime.rememberSavedStateNavEntryDecorator
-import androidx.navigation3.ui.NavDisplay
-import androidx.navigation3.ui.rememberSceneSetupNavEntryDecorator
-import com.google.accompanist.drawablepainter.rememberDrawablePainter
 import com.programmersbox.datastore.DataStoreHandling
-import com.programmersbox.datastore.NewSettingsHandling
-import com.programmersbox.datastore.rememberFloatingNavigation
 import com.programmersbox.favoritesdatabase.ItemDao
 import com.programmersbox.favoritesdatabase.SourceOrder
 import com.programmersbox.kmpmodels.SourceRepository
-import com.programmersbox.kmpuiviews.appVersion
 import com.programmersbox.kmpuiviews.domain.AppUpdate
 import com.programmersbox.kmpuiviews.domain.AppUpdateCheck
 import com.programmersbox.kmpuiviews.presentation.Screen
-import com.programmersbox.kmpuiviews.presentation.components.HazeScaffold
-import com.programmersbox.kmpuiviews.presentation.components.ScreenBottomItem
-import com.programmersbox.kmpuiviews.presentation.navactions.NavigationActions
 import com.programmersbox.kmpuiviews.repository.ChangingSettingsRepository
 import com.programmersbox.kmpuiviews.repository.CurrentSourceRepository
-import com.programmersbox.kmpuiviews.utils.ChromeCustomTabsNavigator
 import com.programmersbox.kmpuiviews.utils.ComposeSettingsDsl
-import com.programmersbox.kmpuiviews.utils.LocalNavActions
-import com.programmersbox.kmpuiviews.utils.LocalNavHostPadding
-import com.programmersbox.kmpuiviews.utils.LocalWindowSizeClass
-import com.programmersbox.kmpuiviews.utils.USE_NAV3
-import com.programmersbox.kmpuiviews.utils.composables.sharedelements.LocalSharedElementScope
-import com.programmersbox.sharedutils.AppLogo
-import com.programmersbox.uiviews.presentation.components.MultipleActions
-import com.programmersbox.uiviews.presentation.components.rememberMultipleBarState
-import com.programmersbox.uiviews.presentation.entryGraph
-import com.programmersbox.uiviews.presentation.navGraph
-import com.programmersbox.uiviews.theme.OtakuMaterialTheme
-import com.programmersbox.uiviews.utils.NotificationLogo
+import com.programmersbox.uiviews.presentation.navigation.HomeNav
 import com.programmersbox.uiviews.utils.currentDetailsUrl
 import com.programmersbox.uiviews.utils.dispatchIo
-import dev.chrisbanes.haze.HazeState
-import dev.chrisbanes.haze.hazeEffect
-import dev.chrisbanes.haze.materials.HazeMaterials
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.koin.android.ext.android.inject
-import kotlin.math.abs
-import kotlin.math.roundToInt
 
 abstract class BaseMainActivity : FragmentActivity() {
 
     protected val genericInfo: GenericInfo by inject()
     private val customPreferences = ComposeSettingsDsl()
         .apply(genericInfo.composeCustomPreferences())
-    private val appLogo: AppLogo by inject()
-    private val notificationLogo: NotificationLogo by inject()
     private val changingSettingsRepository: ChangingSettingsRepository by inject()
     private val itemDao: ItemDao by inject()
     private val dataStoreHandling: DataStoreHandling by inject()
     private val appUpdateCheck: AppUpdateCheck by inject()
-
-    protected lateinit var navController: NavHostController
-
-    protected fun isNavInitialized() = ::navController.isInitialized
-
-    private val settingsHandling: NewSettingsHandling by inject()
-
     private val sourceRepository by inject<SourceRepository>()
     private val currentSourceRepository by inject<CurrentSourceRepository>()
 
@@ -194,12 +49,6 @@ abstract class BaseMainActivity : FragmentActivity() {
     @Composable
     protected open fun BottomBarAdditions() = Unit
 
-    private var notificationCount by mutableIntStateOf(0)
-
-    @OptIn(
-        ExperimentalMaterial3Api::class,
-        ExperimentalMaterial3WindowSizeClassApi::class, ExperimentalSharedTransitionApi::class, ExperimentalMaterial3AdaptiveApi::class
-    )
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setup()
@@ -227,585 +76,18 @@ abstract class BaseMainActivity : FragmentActivity() {
         }
 
         setContent {
-            navController = rememberNavController(
-                remember { ChromeCustomTabsNavigator(this) }
-            )
-
-            val windowSize = calculateWindowSizeClass(activity = this@BaseMainActivity)
-
-            val isAmoledMode by settingsHandling.rememberIsAmoledMode()
-
-            val showBlur by settingsHandling.rememberShowBlur()
-
-            val backStack = rememberNavBackStack(startDestination)
-
-            CompositionLocalProvider(
-                LocalWindowSizeClass provides windowSize
-            ) {
-                OtakuMaterialTheme(
-                    navController = navController,
-                    navBackStack = backStack,
-                    genericInfo = genericInfo,
-                    settingsHandling = settingsHandling,
-                ) {
-                    AskForNotificationPermissions()
-
-                    val showAllItem by settingsHandling.rememberShowAll()
-                    val middleNavItem by settingsHandling.rememberMiddleNavigationAction()
-                    val multipleActions by settingsHandling.rememberMiddleMultipleActions()
-
-                    val navType = when (windowSize.widthSizeClass) {
-                        WindowWidthSizeClass.Expanded -> NavigationBarType.Rail
-                        else -> NavigationBarType.Bottom
-                    }
-
-                    val navBackStackEntry by navController.currentBackStackEntryAsState()
-                    val currentDestination = navBackStackEntry?.destination
-
-                    val showNavBar by changingSettingsRepository.showNavBar.collectAsStateWithLifecycle(true)
-                    val floatingNavigation by rememberFloatingNavigation()
-                    val hazeState = remember { HazeState() }
-
-                    val navigationActions = LocalNavActions.current
-
-                    //val bottomAppBarScrollBehavior = BottomAppBarDefaults.exitAlwaysScrollBehavior()
-
-                    genericInfo.DialogSetups()
-
-                    SharedTransitionLayout {
-                        Row(Modifier.fillMaxSize()) {
-                            Rail(
-                                navActions = navigationActions,
-                                showNavBar = showNavBar,
-                                navType = navType,
-                                showAllItem = showAllItem,
-                                currentDestination = currentDestination
-                            )
-
-                            HazeScaffold(
-                                hazeState = hazeState,
-                                bottomBar = {
-                                    if (!floatingNavigation) {
-                                        BottomNav(
-                                            navActions = navigationActions,
-                                            showNavBar = showNavBar,
-                                            navType = navType,
-                                            currentDestination = currentDestination,
-                                            showBlur = showBlur,
-                                            isAmoledMode = isAmoledMode,
-                                            middleNavItem = middleNavItem,
-                                            multipleActions = multipleActions,
-                                            modifier = Modifier.renderInSharedTransitionScopeOverlay()
-                                        )
-                                    } else {
-                                        HomeNavigationBar(
-                                            showNavBar = showNavBar,
-                                            navType = navType,
-                                            currentDestination = currentDestination,
-                                            showBlur = showBlur,
-                                            isAmoledMode = isAmoledMode,
-                                            middleNavItem = middleNavItem,
-                                            //scrollBehavior = null,
-                                            multipleActions = multipleActions,
-                                            navActions = navigationActions,
-                                            modifier = Modifier
-                                                .padding(horizontal = 24.dp)
-                                                .windowInsetsPadding(WindowInsets.navigationBars)
-                                                .clip(MaterialTheme.shapes.extraLarge)
-                                                .hazeEffect(
-                                                    state = hazeState,
-                                                    style = HazeMaterials.ultraThin(),
-                                                )
-                                                .fillMaxWidth()
-                                                .renderInSharedTransitionScopeOverlay()
-                                        )
-                                    }
-                                },
-                                contentWindowInsets = WindowInsets(0.dp),
-                                //blurBottomBar = showBlur && !floatingNavigation,
-                            ) { innerPadding ->
-                                CompositionLocalProvider(
-                                    LocalNavHostPadding provides innerPadding,
-                                    LocalSharedElementScope provides this@SharedTransitionLayout,
-                                    //LocalHazeState provides hazeState,
-                                    //For later maybe
-                                    //LocalBottomAppBarScrollBehavior provides bottomAppBarScrollBehavior
-                                ) {
-                                    //TODO: Maybe put in a little thing to switch between these two?
-                                    if (USE_NAV3) {
-                                        Nav3(
-                                            backStack = backStack,
-                                            navigationActions = navigationActions,
-                                            genericInfo = genericInfo,
-                                            windowSize = windowSize,
-                                        )
-                                    } else {
-                                        Nav2(
-                                            startDestination = startDestination,
-                                            windowSize = windowSize,
-                                            genericInfo = genericInfo,
-                                            navigationActions = navigationActions,
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @OptIn(ExperimentalMaterial3AdaptiveApi::class)
-    @Composable
-    private fun Nav3(
-        backStack: NavBackStack,
-        navigationActions: NavigationActions,
-        genericInfo: GenericInfo,
-        windowSize: WindowSizeClass,
-    ) {
-        NavDisplay(
-            backStack = backStack,
-            //onBack = { backStack.removeLastOrNull() },
-            sceneStrategy = rememberListDetailSceneStrategy(),
-            onBack = { count ->
-                repeat(count) {
-                    if (backStack.isNotEmpty()) {
-                        backStack.removeLastOrNull()
-                    }
-                }
-            },
-            entryDecorators = listOf(
-                rememberSceneSetupNavEntryDecorator(),
-                rememberSavedStateNavEntryDecorator(),
-                rememberViewModelStoreNavEntryDecorator()
-            ),
-            entryProvider = entryGraph(
+            HomeNav(
+                activity = this,
+                startDestination = startDestination,
                 customPreferences = customPreferences,
-                notificationLogo = notificationLogo,
-                windowSize = windowSize,
-                navigationActions = navigationActions,
-                genericInfo = genericInfo
-            ),
-            transitionSpec = {
-                // Slide in from right when navigating forward
-                slideInHorizontally(initialOffsetX = { it }) togetherWith
-                        slideOutHorizontally(targetOffsetX = { -it })
-            },
-            popTransitionSpec = {
-                // Slide in from left when navigating back
-                slideInHorizontally(initialOffsetX = { -it }) togetherWith
-                        slideOutHorizontally(targetOffsetX = { it })
-            },
-            predictivePopTransitionSpec = {
-                // Slide in from left when navigating back
-                slideInHorizontally(initialOffsetX = { -it }) togetherWith
-                        slideOutHorizontally(targetOffsetX = { it })
-            },
-            modifier = Modifier.fillMaxSize()
-        )
-    }
-
-    @Composable
-    private fun Nav2(
-        startDestination: Screen,
-        windowSize: WindowSizeClass,
-        genericInfo: GenericInfo,
-        navigationActions: NavigationActions,
-    ) {
-        NavHost(
-            navController = navController,
-            startDestination = startDestination,
-            modifier = Modifier.fillMaxSize()
-        ) { navGraph(customPreferences, windowSize, genericInfo, navigationActions, notificationLogo) }
-    }
-
-    @Composable
-    fun FloatingNavigationBar(
-        modifier: Modifier = Modifier,
-        shape: Shape = MaterialTheme.shapes.extraLarge,
-        containerColor: Color = NavigationBarDefaults.containerColor,
-        contentColor: Color = MaterialTheme.colorScheme.contentColorFor(containerColor),
-        tonalElevation: Dp = NavigationBarDefaults.Elevation,
-        content: @Composable RowScope.() -> Unit,
-    ) {
-        Surface(
-            color = containerColor,
-            contentColor = contentColor,
-            tonalElevation = tonalElevation,
-            shape = shape,
-            border = BorderStroke(
-                width = 0.5.dp,
-                brush = Brush.verticalGradient(
-                    colors = listOf(
-                        MaterialTheme.colorScheme.surfaceVariant,
-                        MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.3f),
-                    ),
-                ),
-            ),
-            modifier = modifier,
-        ) {
-            Row(
-                modifier = Modifier
-                    .padding(horizontal = 8.dp)
-                    .fillMaxWidth()
-                    .height(80.dp)
-                    .selectableGroup(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                content = content,
+                bottomBarAdditions = { BottomBarAdditions() }
             )
         }
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
-    @Composable
-    private fun HomeNavigationBar(
-        showNavBar: Boolean,
-        navActions: NavigationActions,
-        navType: NavigationBarType,
-        currentDestination: NavDestination?,
-        showBlur: Boolean,
-        isAmoledMode: Boolean,
-        middleNavItem: com.programmersbox.datastore.MiddleNavigationAction,
-        modifier: Modifier = Modifier,
-        scrollBehavior: BottomAppBarScrollBehavior? = null,
-        multipleActions: com.programmersbox.datastore.MiddleMultipleActions?,
-    ) {
-        val scope = rememberCoroutineScope()
-        AnimatedVisibility(
-            visible = showNavBar && navType == NavigationBarType.Bottom,
-            enter = slideInVertically { it / 2 } + expandVertically() + fadeIn(),
-            exit = slideOutVertically { it / 2 } + shrinkVertically() + fadeOut(),
-        ) {
-            val appBarDragModifier = if (scrollBehavior != null && !scrollBehavior.isPinned) {
-                Modifier.draggable(
-                    orientation = Orientation.Vertical,
-                    state =
-                        rememberDraggableState { delta -> scrollBehavior.state.heightOffset -= delta },
-                    onDragStopped = { velocity ->
-                        settleAppBarBottom(
-                            scrollBehavior.state,
-                            velocity,
-                            scrollBehavior.flingAnimationSpec,
-                            scrollBehavior.snapAnimationSpec
-                        )
-                    }
-                )
-            } else {
-                Modifier
-            }
-
-            val multipleBarState = rememberMultipleBarState()
-
-            Box(modifier = Modifier.fillMaxWidth()) {
-                FloatingNavigationBar(
-                    containerColor = when {
-                        showBlur -> Color.Transparent
-                        isAmoledMode -> MaterialTheme.colorScheme.surface
-                        else -> NavigationBarDefaults.containerColor
-                    },
-                    modifier = modifier
-                        .layout { measurable, constraints ->
-                            // Sets the app bar's height offset to collapse the entire bar's height when
-                            // content
-                            // is scrolled.
-                            scrollBehavior?.state?.heightOffsetLimit = -80.dp.toPx()
-
-                            val placeable = measurable.measure(constraints)
-                            val height = placeable.height + (scrollBehavior?.state?.heightOffset ?: 0f)
-                            layout(placeable.width, height.roundToInt()) { placeable.place(0, 0) }
-                        }
-                        .then(appBarDragModifier),
-                ) {
-                    val colors = NavigationBarItemDefaults.colors(
-                        selectedIconColor = MaterialTheme.colorScheme.onPrimaryContainer,
-                        selectedTextColor = MaterialTheme.colorScheme.onPrimaryContainer,
-                        unselectedIconColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f),
-                        unselectedTextColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f),
-                    )
-
-                    @Composable
-                    fun ScreenBottomItem(
-                        screen: Screen,
-                        icon: ImageVector,
-                        label: Int,
-                        badge: @Composable BoxScope.() -> Unit = {},
-                    ) {
-                        NavigationBarItem(
-                            icon = { BadgedBox(badge = badge) { Icon(icon, null) } },
-                            label = { Text(stringResource(label)) },
-                            selected = currentDestination.isTopLevelDestinationInHierarchy(screen),
-                            colors = colors,
-                            onClick = { navActions.homeScreenNavigate(screen) }
-                        )
-                    }
-
-                    ScreenBottomItem(
-                        screen = Screen.RecentScreen,
-                        icon = if (currentDestination.isTopLevelDestinationInHierarchy(Screen.RecentScreen)) Icons.Default.History else Icons.Outlined.History,
-                        label = R.string.recent
-                    )
-
-                    middleNavItem.ScreenBottomItem(
-                        rowScope = this,
-                        currentDestination = currentDestination,
-                        colors = colors,
-                        multipleClick = {
-                            scope.launch {
-                                if (multipleBarState.showHorizontalBar) {
-                                    multipleBarState.hide()
-                                } else {
-                                    multipleBarState.show()
-                                }
-                            }
-                        },
-                        onClick = { navActions.homeScreenNavigate(it.screen) }
-                    )
-
-                    ScreenBottomItem(
-                        screen = Screen.Settings,
-                        icon = if (currentDestination.isTopLevelDestinationInHierarchy(Screen.Settings)) Icons.Default.Settings else Icons.Outlined.Settings,
-                        label = R.string.settings,
-                        badge = { if (updateCheck()) Badge { Text("") } }
-                    )
-                }
-
-                multipleActions?.let {
-                    MultipleActions(
-                        state = multipleBarState,
-                        middleNavItem = middleNavItem,
-                        multipleActions = it,
-                        currentDestination = currentDestination,
-                        navigationActions = navActions
-                    )
-                }
-            }
-        }
-    }
-
-    @Composable
-    private fun BottomNav(
-        navActions: NavigationActions,
-        showNavBar: Boolean,
-        navType: NavigationBarType,
-        currentDestination: NavDestination?,
-        showBlur: Boolean,
-        isAmoledMode: Boolean,
-        middleNavItem: com.programmersbox.datastore.MiddleNavigationAction,
-        multipleActions: com.programmersbox.datastore.MiddleMultipleActions?,
-        modifier: Modifier = Modifier,
-    ) {
-        val scope = rememberCoroutineScope()
-        val multipleBarState = rememberMultipleBarState()
-        Box(modifier = modifier) {
-            Column {
-                BottomBarAdditions()
-                AnimatedVisibility(
-                    visible = showNavBar && navType == NavigationBarType.Bottom,
-                    enter = slideInVertically { it / 2 } + expandVertically() + fadeIn(),
-                    exit = slideOutVertically { it / 2 } + shrinkVertically() + fadeOut(),
-                ) {
-                    NavigationBar(
-                        containerColor = when {
-                            showBlur -> Color.Transparent
-                            isAmoledMode -> MaterialTheme.colorScheme.surface
-                            else -> NavigationBarDefaults.containerColor
-                        }
-                    ) {
-
-                        @Composable
-                        fun ScreenBottomItem(
-                            screen: Screen,
-                            icon: ImageVector,
-                            label: Int,
-                            badge: @Composable BoxScope.() -> Unit = {},
-                        ) {
-                            NavigationBarItem(
-                                icon = { BadgedBox(badge = badge) { Icon(icon, null) } },
-                                label = { Text(stringResource(label)) },
-                                selected = currentDestination.isTopLevelDestinationInHierarchy(screen),
-                                onClick = { navActions.homeScreenNavigate(screen) }
-                            )
-                        }
-
-                        ScreenBottomItem(
-                            screen = Screen.RecentScreen,
-                            icon = if (currentDestination.isTopLevelDestinationInHierarchy(Screen.RecentScreen)) Icons.Default.History else Icons.Outlined.History,
-                            label = R.string.recent
-                        )
-
-                        middleNavItem.ScreenBottomItem(
-                            rowScope = this,
-                            currentDestination = currentDestination,
-                            multipleClick = {
-                                scope.launch {
-                                    if (multipleBarState.showHorizontalBar) {
-                                        multipleBarState.hide()
-                                    } else {
-                                        multipleBarState.show()
-                                    }
-                                }
-                            },
-                            onClick = { navActions.homeScreenNavigate(it.screen) }
-                        )
-
-                        ScreenBottomItem(
-                            screen = Screen.Settings,
-                            icon = if (currentDestination.isTopLevelDestinationInHierarchy(Screen.Settings)) Icons.Default.Settings else Icons.Outlined.Settings,
-                            label = R.string.settings,
-                            badge = { if (updateCheck()) Badge { Text("") } }
-                        )
-                    }
-                }
-            }
-
-            multipleActions?.let {
-                MultipleActions(
-                    state = multipleBarState,
-                    middleNavItem = middleNavItem,
-                    multipleActions = it,
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-            }
-        }
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    private fun Rail(
-        navActions: NavigationActions,
-        showNavBar: Boolean,
-        navType: NavigationBarType,
-        showAllItem: Boolean,
-        currentDestination: NavDestination?,
-    ) {
-        AnimatedVisibility(
-            visible = showNavBar && navType == NavigationBarType.Rail,
-            enter = slideInHorizontally { it / 2 } + expandHorizontally() + fadeIn(),
-            exit = slideOutHorizontally { it / 2 } + shrinkHorizontally() + fadeOut(),
-        ) {
-            NavigationRail(
-                header = {
-                    Image(
-                        rememberDrawablePainter(drawable = appLogo.logo),
-                        null,
-                    )
-                },
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .verticalScroll(rememberScrollState())
-            ) {
-                NavigationRailItem(
-                    imageVector = Icons.Default.History,
-                    label = stringResource(R.string.recent),
-                    screen = Screen.RecentScreen,
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-
-                AnimatedVisibility(visible = showAllItem) {
-                    NavigationRailItem(
-                        imageVector = Icons.Default.BrowseGallery,
-                        label = stringResource(R.string.all),
-                        screen = Screen.AllScreen,
-                        currentDestination = currentDestination,
-                        navigationActions = navActions
-                    )
-                }
-
-                AnimatedVisibility(visible = notificationCount > 0) {
-                    NavigationRailItem(
-                        imageVector = Icons.Default.Notifications,
-                        label = stringResource(R.string.notifications),
-                        screen = Screen.NotificationScreen,
-                        currentDestination = currentDestination,
-                        navigationActions = navActions
-                    )
-                }
-
-                NavigationRailItem(
-                    imageVector = Icons.AutoMirrored.Default.List,
-                    label = stringResource(R.string.custom_lists_title),
-                    screen = Screen.CustomListScreen,
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-
-                NavigationRailItem(
-                    imageVector = Icons.Default.Search,
-                    label = stringResource(R.string.global_search),
-                    screen = Screen.GlobalSearchScreen(),
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-
-                NavigationRailItem(
-                    imageVector = Icons.Default.Star,
-                    label = stringResource(R.string.viewFavoritesMenu),
-                    screen = Screen.FavoriteScreen,
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-
-                NavigationRailItem(
-                    imageVector = Icons.Default.Extension,
-                    label = stringResource(R.string.extensions),
-                    screen = Screen.ExtensionListScreen,
-                    currentDestination = currentDestination,
-                    navigationActions = navActions
-                )
-
-                NavigationRailItem(
-                    icon = {
-                        BadgedBox(
-                            badge = { if (updateCheck()) Badge { Text("") } }
-                        ) { Icon(Icons.Default.Settings, null) }
-                    },
-                    label = { Text(stringResource(R.string.settings)) },
-                    selected = currentDestination.isTopLevelDestinationInHierarchy(Screen.Settings),
-                    onClick = { navActions.homeScreenNavigate(Screen.Settings) }
-                )
-            }
-        }
-    }
-
-    @Composable
-    private fun NavigationRailItem(
-        imageVector: ImageVector,
-        label: String,
-        screen: Screen,
-        currentDestination: NavDestination?,
-        navigationActions: NavigationActions,
-        onClick: () -> Unit = { navigationActions.homeScreenNavigate(screen) },
-    ) {
-        NavigationRailItem(
-            icon = { Icon(imageVector, null) },
-            label = { Text(label) },
-            selected = currentDestination.isTopLevelDestinationInHierarchy(screen.route),
-            onClick = onClick
-        )
-    }
-
-    @Composable
-    fun AskForNotificationPermissions() {
-        if (Build.VERSION.SDK_INT >= 33) {
-            val permissionRequest = rememberLauncherForActivityResult(ActivityResultContracts.RequestPermission()) {}
-            LaunchedEffect(Unit) { permissionRequest.launch(Manifest.permission.POST_NOTIFICATIONS) }
-        }
-    }
-
-    @Composable
-    fun updateCheck(): Boolean {
-        val appUpdate by appUpdateCheck.updateAppCheck.collectAsStateWithLifecycle(null)
-        return AppUpdate.checkForUpdate(
-            appVersion(),
-            appUpdate?.updateRealVersion.orEmpty()
-        )
     }
 
     private fun setup() {
-        sourceRepository.sources
+        sourceRepository
+            .sources
             .onEach {
                 it.forEachIndexed { index, sourceInformation ->
                     itemDao.insertSourceOrder(
@@ -826,7 +108,7 @@ abstract class BaseMainActivity : FragmentActivity() {
                 if (it == null) {
                     sourceRepository
                         .list
-                        .filter { it.catalog == null }
+                        .filter { c -> c.catalog == null }
                         .randomOrNull()
                 } else {
                     sourceRepository.toSourceByApiServiceName(it)
@@ -840,81 +122,10 @@ abstract class BaseMainActivity : FragmentActivity() {
             .dispatchIo()
             .onEach(appUpdateCheck.updateAppCheck::emit)
             .launchIn(lifecycleScope)
-
-        itemDao
-            .getAllNotificationCount()
-            .onEach { notificationCount = it }
-            .launchIn(lifecycleScope)
     }
-
-    private fun NavDestination?.isTopLevelDestinationInHierarchy(destination: Screen) = isTopLevelDestinationInHierarchy(destination.route)
-
-    private fun NavDestination?.isTopLevelDestinationInHierarchy(destination: String) = this?.hierarchy?.any {
-        it.route?.contains(destination, true) == true
-    } == true
 
     override fun onProvideAssistContent(outContent: AssistContent?) {
         super.onProvideAssistContent(outContent)
         outContent?.webUri = currentDetailsUrl.toUri()
-    }
-
-    override fun onNewIntent(intent: Intent) {
-        super.onNewIntent(intent)
-        if (isNavInitialized()) navController.handleDeepLink(intent)
-    }
-
-    enum class NavigationBarType { Rail, Bottom }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    private suspend fun settleAppBarBottom(
-        state: BottomAppBarState,
-        velocity: Float,
-        flingAnimationSpec: DecayAnimationSpec<Float>?,
-        snapAnimationSpec: AnimationSpec<Float>?,
-    ): Velocity {
-        // Check if the app bar is completely collapsed/expanded. If so, no need to settle the app bar,
-        // and just return Zero Velocity.
-        // Note that we don't check for 0f due to float precision with the collapsedFraction
-        // calculation.
-        if (state.collapsedFraction < 0.01f || state.collapsedFraction == 1f) {
-            return Velocity.Zero
-        }
-        var remainingVelocity = velocity
-        // In case there is an initial velocity that was left after a previous user fling, animate to
-        // continue the motion to expand or collapse the app bar.
-        if (flingAnimationSpec != null && abs(velocity) > 1f) {
-            var lastValue = 0f
-            AnimationState(
-                initialValue = 0f,
-                initialVelocity = velocity,
-            )
-                .animateDecay(flingAnimationSpec) {
-                    val delta = value - lastValue
-                    val initialHeightOffset = state.heightOffset
-                    state.heightOffset = initialHeightOffset + delta
-                    val consumed = abs(initialHeightOffset - state.heightOffset)
-                    lastValue = value
-                    remainingVelocity = this.velocity
-                    // avoid rounding errors and stop if anything is unconsumed
-                    if (abs(delta - consumed) > 0.5f) this.cancelAnimation()
-                }
-        }
-        // Snap if animation specs were provided.
-        if (snapAnimationSpec != null) {
-            if (state.heightOffset < 0 && state.heightOffset > state.heightOffsetLimit) {
-                AnimationState(initialValue = state.heightOffset).animateTo(
-                    if (state.collapsedFraction < 0.5f) {
-                        0f
-                    } else {
-                        state.heightOffsetLimit
-                    },
-                    animationSpec = snapAnimationSpec
-                ) {
-                    state.heightOffset = value
-                }
-            }
-        }
-
-        return Velocity(0f, remainingVelocity)
     }
 }
