+++ gradle/libs.versions.toml
@@ -35,7 +35,7 @@ roomVersion = "2.7.1"
 navVersion = "2.9.0"
 koin-bom = "4.0.4"
 coil = "3.2.0"
-landscapist = "2.4.8"
+landscapist = "2.4.9"
 lifecycle = "2.9.0"
 jetpack = "1.6.0-alpha03"
 jetpackCompiler = "1.5.11"
